[Feat] [FE-57] icon 분리 및 컴포넌트 삭제
